# services/core/core/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Core API
  description: API centrale servant de façade pour plusieurs autres API.
  version: 0.1
  contact:
    name: Support API
    email: support@example.com
servers:
  - url: http://localhost:8080
    description: Serveur de développement local
...
paths:
  /messages/student/{student}:
    get:
      summary: Get all messages for a student
      tags:
        - Messages
      parameters:
        - name: student
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /messages:
    post:
      summary: Create a new message
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid input
  /messages/{id}:
    get:
      summary: Get a message by ObjectId
      tags:
        - Messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found
    put:
      summary: Update a message by ObjectId
      tags:
        - Messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: Message updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid input
        '404':
          description: Message not found
    delete:
      summary: Delete a message by ObjectId
      tags:
        - Messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Message deleted
        '404':
          description: Message not found
  /auth/reset-password/{username}:
    put:
      tags:
        - auth-controller
      operationId: resetPassword
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /auth/register:
    post:
      tags:
        - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Already registered
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicError'
  /auth/login:
    post:
      tags:
        - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid credentials
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicError'
  /auth/me:
    get:
      tags:
        - auth-controller
      operationId: validateToken
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Invalid token
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicError'
        '403':
          description: Banned token
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicError'
  /auth/unregister:
    delete:
      tags:
        - auth-controller
      operationId: unregister
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicResponse'
        '400':
          description: Invalid credentials
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BasicError'
  /auth/logout:
    delete:
      tags:
        - auth-controller
      operationId: logout
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /students/{id}:
    get:
      tags:
        - student-controller
      summary: Récupérer un étudiant par ID
      description: Retourne les informations d'un étudiant à partir de son ID.
      operationId: getStudentById
      parameters:
        - name: id
          in: path
          required: true
          description: ID unique de l'étudiant
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Étudiant récupéré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    put:
      tags:
        - student-controller
      summary: Mettre à jour un étudiant
      description: Modifie les informations d'un étudiant existant.
      operationId: updateStudent
      parameters:
        - name: id
          in: path
          required: true
          description: ID unique de l'étudiant à mettre à jour
          schema:
            type: integer
            format: int64
      requestBody:
        description: Informations mises à jour de l'étudiant
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      responses:
        '200':
          description: Étudiant mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    delete:
      tags:
        - student-controller
      summary: Supprimer un étudiant
      description: Supprime un étudiant de la base de données à partir de son ID.
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          required: true
          description: ID unique de l'étudiant à supprimer
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Étudiant supprimé avec succès
  /students:
    get:
      tags:
        - student-controller
      summary: Récupérer la liste des étudiants
      description: Retourne la liste complète des étudiants enregistrés dans la base de données.
      operationId: getAllStudents
      parameters:
        - in: query
          name: formation
          schema:
            type: integer
            format: int64
          description: Filtre les étudiants par formation
      responses:
        '200':
          description: Liste des étudiants récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      tags:
        - student-controller
      summary: Ajouter un nouvel étudiant
      description: Ajoute un nouvel étudiant à la base de données.
      operationId: createStudent
      requestBody:
        description: Données du nouvel étudiant à enregistrer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      responses:
        '201':
          description: Étudiant créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
...
components:
  schemas:
    Message:
      type: object
      properties:
        _id:
          type: string
          description: Unique ObjectId of the message
          example: 5f92cdce0cf217478ba93563
        text:
          type: string
          description: Message content
          example: Votre inscription au Master TIIL-A a été accepter.
        readed:
          type: boolean
          description: Message read status
          example: false
        student:
          type: integer
          description: ID of the student associated with the message
          example: 26
      required:
        - text
        - student
        - readed
    BasicResponse:
      type: object
      properties:
        message:
          type: string
    BasicError:
        type: object
        properties:
          error:
            type: string
    RegisterRequest:
        required:
          - password
          - student
          - username
        type: object
        properties:
          username:
            type: string
          password:
            type: string
          student:
            type: integer
            format: int64
    AuthRequest:
        required:
          - password
          - username
        type: object
        properties:
          username:
            type: string
          password:
            type: string
    TokenResponse:
        type: object
        properties:
          message:
            type: string
          token:
            type: string
    MeResponse:
        type: object
        properties:
          ok:
            type: boolean
          studentNumber:
            type: integer
            format: int64
    Student:
      required:
        - adress
        - firstname
        - surname
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identifiant unique de l'étudiant
        academicYearRegistered:
          type: boolean
          description: Indique si l'étudiant est inscrit pour l'année académique en cours
        coursesId:
          type: array
          items:
            type: integer
            format: int64
          description: Liste des identifiants des cours suivis par l'étudiant
        academicYearId:
          type: integer
          format: int64
          description: Identifiant de l'année académique
        surname:
          type: string
          description: Nom de famille de l'étudiant
        firstname:
          type: string
          description: Prénom de l'étudiant
        adress:
          type: string
          description: Adresse de l'étudiant
        pw:
          minimum: 1
          type: integer
          format: int32
          description: ID du groupe de TP de l'étudiant (PW)
        dw:
          minimum: 1
          type: integer
          format: int32
          description: ID du groupe de TD de l'étudiant (DW)
...